+++ NEWS.md
@@ -1,7 +1,7 @@
 # socialranking 0.1.0-9004 (change this part until 0.1.1 is released)
 
 * Added parameter `equivalenceClasses` to `newPowerRelation`
-* `equivalenceClassIndex` returns vector instead of single number
+* Added `equivalenceClassIndexes` to return multiple eq numbers if power relation is cyclic
 * Fix lexcel equals operator
 
 # socialranking 0.1.0 (2022-03-28)